Rename compilation buffers to command name
Adds the buffer name function that renames the compilation buffer name
to include the name of the command being ran. This makes it so you can
have multiple commands running at the same time. If, for example, a
command stays running (a watch command, for instance), you would not
be prevented from running others while that one is active, unless you
want to run the same one, of course.

It does not take into account where the command came from, so there
could still be collissions with commands from different projects that
have the same name.