Error on unknown format specifiers
This behavior is better than ignore. Suppose you write a command with
a currently non-existent specifier like %z. With 'ignore that is
passed verbatim to the runner. If later on there's a Conner update
that adds a specifier to %z, then the behavior of your command
suddenly changes. It's better to enforce that if you want to pass %z
to the command, then you have to escape it with %%z. Otherwise, an
error is raised.